requirements.txt
fastapi>=0.85
uvicorn[standard]>=0.18
firebase-admin>=6.0
pydantic>=1.10
google-cloud-firestore>=2.0
joblib
python-dotenv>=0.21.0
fastapi


pandas
numpy
scikit-learn

python-multipart

flask
flask-cors
firebase-admin
pydantic
python-multipart

pillow
pytesseract
pdf2image
opencv-python
numpy
python-dotenv
requests
scikit-learn
joblib
pandas
sentence-transformers
faiss-cpu

fastapi
uvicorn

fastapi>=0.85
uvicorn[standard]>=0.18
firebase-admin>=6.0
pydantic>=1.10
google-cloud-firestore>=2.0
joblib
python-dotenv>=0.21.0
